dw3000: Don't attempt reading reset GPIO
gpio_direction_input should setup the GPIO as
input so that we can call gpio_get_value on it.
But something goes horribly wrong and gpio_get_value
always returns the logical value 0

The addition to is only found in the 13 stack
and it seems we aren't alone reverting this change,
Google removed it from their gs101/gs201 platform.

Furthermore, the DW3720 user manual states that the
reset pin should never be driven high, so this is
clearly incorrect. The intention seems to be to wait
for the UWB chipset to drive the reset line high,
indicating that it's finished cold boot, but this
isn't necessary. And using -EPROBE_DEFER is clearly
the wrong mechanism to use here; a poll loop would
be more appropriate.

Change-Id: Iffc20ed3b84b2b25526819d87423810f6a9a9a48
Signed-off-by: Luca Stefani <luca@osomprivacy.com>
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>