dw3000: Always enable PLL reference clock if specified
In the IDLE_PLL state, the system clock source is switched from FAST_RC to
the internal PLL clock, which is clocked by an external reference clock at
38.4 MHz.

The external reference clock must be enabled in order for the PLL clock to
lock and thus transition the chipset to the IDLE_PLL state.

Parse this clock as "rf_clk" from the device tree and always enable it in
order to make the chipset functional.

Change-Id: I38c626b6df8ec8e42a852ee02ebf6b862294389b
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>