dw3000: Support triggering a wake using the wakeup pin
The secondary SPI chip select pin (SPI2CS) can be utilized to trigger a
wake from deep sleep if the secondary SPI bus is unused. This wakeup pin is
active high and should be driven as such for 500 us in order to trigger a
wake event and thus kick the chipset out of deep sleep.

This is helpful for when the primary wake mechanism, driving the master SPI
chip select pin (SPI1CSn) low for 500 us by triggering a dummy SPI read,
doesn't work to trigger a wake event.

This requires configuring the WAKE_WUP field to 1 in the AON_CFG register,
to permit using the wakeup pin as a wake-up signal.

Change-Id: Ibd0b3d143a69e9014076ac014cf1f2857f9548e7
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>